Entry class refactor
